Updated from MicroManager v1
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@33 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/Yokogawa/CSU22.h
@@ -180,7 +180,7 @@ private:
    bool driveSpeedBusy_;
    unsigned numPos_; 
    std::string name_;
-   long min_, max_, current_, requested_;
+   int min_, max_, current_, requested_;
 };
